feat(api-types): move domain entities into the crate
Move all entity structs + co-located status enums (Invoice, Payment,
Wallet, Balance, Contact, LnAddress, BtcAddress/BtcOutput, ApiKey,
LnUrlSuccessAction/LnUrlCallback, *Status) into swissknife-api-types;
domain modules re-export them. Filters + OrderDirection stay in swissknife
(sea_orm coupling). Bolt11Invoice->Invoice becomes a free function (orphan);
api-types pulls the lightweight nostr core for PublicKey.
