+++ crates/swissknife-api-types/src/bitcoin.rs
@@ -1,0 +1,63 @@
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use strum_macros::{Display, EnumString};
+use utoipa::ToSchema;
+use uuid::Uuid;
+
+use crate::InvoiceStatus;
+
+#[derive(Clone, Debug)]
+pub struct BtcAddress {
+    pub id: Uuid,
+    pub wallet_id: Uuid,
+    pub address: String,
+    pub used: bool,
+    pub address_type: BtcAddressType,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: Option<DateTime<Utc>>,
+}
+
+#[derive(Clone, Debug, Copy, Deserialize, Serialize, EnumString, Display, PartialEq, Eq, Default, ToSchema)]
+#[serde(rename_all = "lowercase")]
+#[strum(serialize_all = "lowercase")]
+pub enum BtcAddressType {
+    P2pkh,
+    P2sh,
+    #[default]
+    P2wpkh,
+    P2tr,
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct BtcOutput {
+    pub id: Uuid,
+    pub outpoint: String,
+    pub txid: String,
+    pub output_index: u32,
+    pub address: String,
+    pub amount_sat: u64,
+    pub status: BtcOutputStatus,
+    pub block_height: Option<u32>,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: Option<DateTime<Utc>>,
+}
+
+#[derive(Clone, Debug, Copy, EnumString, Deserialize, Serialize, Display, PartialEq, Eq, Default, ToSchema)]
+pub enum BtcOutputStatus {
+    #[default]
+    Unconfirmed,
+    Confirmed,
+    Spent,
+    Immature,
+}
+
+impl From<BtcOutputStatus> for InvoiceStatus {
+    fn from(status: BtcOutputStatus) -> Self {
+        match status {
+            BtcOutputStatus::Unconfirmed => InvoiceStatus::Pending,
+            BtcOutputStatus::Confirmed => InvoiceStatus::Settled,
+            BtcOutputStatus::Spent => InvoiceStatus::Settled,
+            BtcOutputStatus::Immature => InvoiceStatus::Pending,
+        }
+    }
+}
