+++ crates/swissknife-api-types/src/lib.rs
@@ -5,10 +5,24 @@
 //! clients/SDKs. They carry their own serde + `utoipa::ToSchema` annotations so
 //! the wire shape lives with the type; behaviour stays in the app's use cases.
 
+mod api_key;
+mod bitcoin;
+mod invoice;
+mod ln_address;
+mod lnurl;
 mod network;
+mod payment;
 mod permission;
 mod transaction;
+mod wallet;
 
+pub use api_key::ApiKey;
+pub use bitcoin::{BtcAddress, BtcAddressType, BtcOutput, BtcOutputStatus};
+pub use invoice::{Invoice, InvoiceStatus, LnInvoice};
+pub use ln_address::LnAddress;
+pub use lnurl::{LnUrlCallback, LnUrlSuccessAction};
 pub use network::BtcNetwork;
+pub use payment::{BtcPayment, InternalPayment, LnPayment, Payment, PaymentStatus};
 pub use permission::Permission;
 pub use transaction::{Currency, Ledger};
+pub use wallet::{Balance, Contact, Wallet, WalletOverview};
