+++ src/domains/lnurl/entities/lnurl.rs
@@ -2,6 +2,8 @@ use nostr_sdk::PublicKey;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
+pub use swissknife_api_types::{LnUrlCallback, LnUrlSuccessAction};
+
 #[derive(Debug, Serialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct LnURLPayRequest {
@@ -72,30 +74,4 @@ pub enum LnUrlPaySuccessAction {
         ciphertext: String,
         iv: String,
     },
-}
-
-#[derive(Debug)]
-pub struct LnUrlCallback {
-    pub pr: String,
-    pub success_action: Option<LnUrlSuccessAction>,
-    pub disposable: Option<bool>,
-    pub routes: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema, Default)]
-pub struct LnUrlSuccessAction {
-    /// Action type. One of `url` or `message`
-    pub tag: String,
-
-    /// Message displayed to the user
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-
-    /// Description
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-
-    /// URL for the user to open on success
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub url: Option<String>,
 }
