+++ src/domains/user/entities/api_key.rs
@@ -1,4 +1,3 @@
-use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 use utoipa::IntoParams;
@@ -6,20 +5,7 @@ use uuid::Uuid;
 
 use crate::application::entities::OrderDirection;
 
-use super::Permission;
-
-#[derive(Clone, Debug, Default)]
-pub struct ApiKey {
-    pub id: Uuid,
-    pub user_id: String,
-    pub name: String,
-    pub key: Option<String>,
-    pub key_hash: Vec<u8>,
-    pub permissions: Vec<Permission>,
-    pub description: Option<String>,
-    pub created_at: DateTime<Utc>,
-    pub expires_at: Option<DateTime<Utc>>,
-}
+pub use swissknife_api_types::ApiKey;
 
 #[serde_as]
 #[derive(Clone, Debug, Deserialize, Serialize, Default, IntoParams)]
