+++ src/domains/wallet/entities/balance.rs
@@ -1,25 +1,1 @@
-use serde::Serialize;
-use utoipa::ToSchema;
-
-#[derive(Debug, Clone, Serialize, Default, ToSchema)]
-pub struct Balance {
-    /// Total amount received
-    #[schema(example = 1000000000)]
-    pub received_msat: u64,
-
-    /// Total amount sent (settled outgoing payments)
-    #[schema(example = 10000000)]
-    pub sent_msat: u64,
-
-    /// Total fees paid
-    #[schema(example = 1000)]
-    pub fees_paid_msat: u64,
-
-    /// Amount reserved for pending outgoing payments
-    #[schema(example = 2000)]
-    pub reserved_msat: u64,
-
-    /// Amount available to spend.
-    #[schema(example = 989999000)]
-    pub available_msat: i64,
-}
+pub use swissknife_api_types::Balance;
